requirements.txt
aiogram~=3.10.0
pillow~=10.4.0
gspread~=6.1.2
oauth2client~=4.1.3
aiofiles~=23.2.1